refactor(projects): consolidate gallery into Projects component and enhance dialog responsiveness
Merge ProjectGallery functionality directly into Projects component:
- Remove separate ProjectGallery component
- Integrate SWR data fetching into Projects
- Add Dialog-based project detail view
- Update font imports (Titillium Web and Saira)
- Enhance dialog sizing for portrait and landscape orientations
- Improve dialog overlay behavior on mobile devices
- Add min/max constraints for dialog dimensions
