+++ src/app/layout.tsx
@@ -1,5 +1,12 @@
 import type { Metadata, Viewport } from "next";
-import { Geist, Geist_Mono, Inter, Merriweather } from "next/font/google";
+import {
+  Geist,
+  Geist_Mono,
+  Inter,
+  Merriweather,
+  Saira,
+  Titillium_Web,
+} from "next/font/google";
 import { ThemeWrapper } from "@/components/ThemeWrapper";
 import "./globals.css";
 
@@ -13,13 +20,13 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-const inter = Inter({
+const inter = Titillium_Web({
   variable: "--font-inter",
   weight: ["400", "600"],
   subsets: ["latin"],
 });
 
-const merriweather = Merriweather({
+const merriweather = Saira({
   variable: "--font-mw",
   weight: ["400", "700"],
   subsets: ["latin"],
