+++ src/components/ui/dialog.tsx
@@ -39,7 +39,7 @@ function DialogOverlay({
     <DialogPrimitive.Overlay
       data-slot="dialog-overlay"
       className={cn(
-        "data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 fixed inset-0 z-50", // bg-black/50
+        "data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 fixed inset-0 z-50 portrait:bg-black/50 md:portrait:bg-black/0", // bg-black/50
         className
       )}
       {...props}
@@ -62,8 +62,10 @@ function DialogContent({
         data-slot="dialog-content"
         className={cn(
           "bg-background data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95",
-          "fixed top-[50%] left-[50%] z-50 grid w-full max-w-[calc(100%-2rem)] translate-x-[-50%] translate-y-[-50%] gap-4 rounded-[25px] border p-4 shadow-lg duration-200",
-          "md:portrait:w-[var(--face-size)] md:portrait:h-[var(--face-size)]",
+          "fixed top-[50%] left-[50%] z-50 grid w-full max-w-[calc(100%-2rem)] translate-x-[-50%] translate-y-[-50%] gap-4 rounded-[25px] border-none p-2 pt-4 md:p-8 shadow-lg duration-200",
+          "min-h-[300px] min-w-[300px]",
+          "md:portrait:w-[var(--face-size)] md:portrait:max-h-[1400px] md:portrait:h-[calc(100vh-200px)]",
+          "md:landscape:w-[calc(100vw-200px)] md:landscape:max-w-[1400px] md:landscape:h-[var(--face-size)]",
           className
         )}
         {...props}
